pyproject.toml
[project]
name = "simsim-tools-server"
version = "0.1.0"
description = ""
authors = [
    {name = "FernandoYu",email = "fyu@fi.uba.ar"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.123.9,<0.124.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pymupdf (>=1.26.6,<2.0.0)"
]

[tool.poetry]
packages = [{include = "simsim_tools_server", from = "src"}]

[tool.poetry.scripts]
main = "simsim_tools_server.main:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
